requirements.txt
streamlit 
pandas
requests
pickle
python-dotenv
scikit-learn
langchain_huggingface
transformers
